+++ common/src/main/java/com/chefmooon/ubesdelight/common/CommonSetup.java
@@ -5,11 +5,20 @@ import com.chefmooon.ubesdelight.common.block.entity.dispenser.DrinkableFeastDis
 import com.chefmooon.ubesdelight.common.block.entity.dispenser.BaseLeafFeastDispenseBehavior;
 import com.chefmooon.ubesdelight.common.registry.UbesDelightItems;
 import com.chefmooon.ubesdelight.common.utility.BuiltInRegistryUtil;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Sets;
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.ComposterBlock;
+
+import java.util.HashMap;
+import java.util.Set;
 
 public class CommonSetup {
     public static void init() {
         registerDispenserBehaviors();
+        registerItemSetAdditions();
     }
 
     public static void registerDispenserBehaviors() {
@@ -53,5 +62,25 @@ public class CommonSetup {
 
         ComposterBlock.COMPOSTABLES.put(BuiltInRegistryUtil.getItem(UbesDelightItems.UBE_CAKE), 1.0f);
         ComposterBlock.COMPOSTABLES.put(BuiltInRegistryUtil.getItem(UbesDelightItems.LECHE_FLAN_FEAST), 1.0f);
+    }
+
+    public static void registerItemSetAdditions() {
+        Set<Item> newWantedItems = Sets.newHashSet(
+                BuiltInRegistryUtil.getItem(UbesDelightItems.UBE),
+                BuiltInRegistryUtil.getItem(UbesDelightItems.GARLIC),
+                BuiltInRegistryUtil.getItem(UbesDelightItems.GINGER),
+                BuiltInRegistryUtil.getItem(UbesDelightItems.LEMONGRASS),
+                BuiltInRegistryUtil.getItem(UbesDelightItems.LEMONGRASS_SEEDS)
+        );
+        newWantedItems.addAll(Villager.WANTED_ITEMS);
+        Villager.WANTED_ITEMS = ImmutableSet.copyOf(newWantedItems);
+
+        HashMap<Item, Integer> newFoodPoints = new HashMap<>();
+        newFoodPoints.put(BuiltInRegistryUtil.getItem(UbesDelightItems.UBE), 1);
+        newFoodPoints.put(BuiltInRegistryUtil.getItem(UbesDelightItems.GARLIC), 1);
+        newFoodPoints.put(BuiltInRegistryUtil.getItem(UbesDelightItems.GINGER), 1);
+        newFoodPoints.put(BuiltInRegistryUtil.getItem(UbesDelightItems.LEMONGRASS), 1);
+        newFoodPoints.putAll(Villager.FOOD_POINTS);
+        Villager.FOOD_POINTS = ImmutableMap.copyOf(newFoodPoints);
     }
 }