+++ fabric/src/main/java/com/chefmooon/ubesdelight/data/fabric/ItemTagGenerator.java
@@ -65,6 +65,12 @@ public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
         getOrCreateTagBuilder(ItemTags.DURABILITY_ENCHANTABLE).addTag(UbesDelightTags.TOOLS_ROLLING_PIN);
         getOrCreateTagBuilder(ItemTags.MINING_LOOT_ENCHANTABLE).addTag(UbesDelightTags.TOOLS_ROLLING_PIN);
         getOrCreateTagBuilder(ItemTags.BREAKS_DECORATED_POTS).addTag(UbesDelightTags.TOOLS_ROLLING_PIN);
+
+        getOrCreateTagBuilder(ItemTags.VILLAGER_PLANTABLE_SEEDS)
+                .add(UbesDelightItemsImpl.UBE)
+                .add(UbesDelightItemsImpl.GARLIC)
+                .add(UbesDelightItemsImpl.GINGER)
+                .add(UbesDelightItemsImpl.LEMONGRASS_SEEDS);
     }
     private void registerCompatibiltyItemTags() {
         // Create Item Tags