+++ changelog.md
@@ -7,6 +7,10 @@ NeoForge and Fabric.
 
 ***
 
+### Added
+
+- Villagers can now consume, harvest, and plant Ube, Garlic, Ginger, and Lemongrass crops
+
 ### Changed
 
 - Match FD 1.2.10 Tag Changes: